+++ Padding.java
@@ -1,7 +1,7 @@
 public class Padding{
 
 public static String input = "";
-public static int[] hexStrings = new int[16]; //for 512 bit strings there will be 64 bytes. there will be 16 hex vals.
+public static long[] hexStrings = new long[16]; //for 512 bit strings there will be 64 bytes. there will be 16 hex vals.
 
 
 
@@ -36,8 +36,8 @@ public static String intToBinary(int int1){
     return str;
 }
 
-public static int binaryStringtoInt(String str){
-  return Integer.parseInt(str,2);
+public static Long binaryStringtoInt(String str){
+  return Long.parseLong(str,2);
 }
 
 public static String zerosPad(int location,int endlength,String initStr){
